Add ability to extract task and run name from protocol name
If the protocol name matches task-([a-zA-z]+)_run([0-9]+)
the matching groups will be used as task & run labels in the
BIDS file name. Otherwise, generic 'task-fmri' rules still apply.

This creates a need to discard runs of a given task that repeated
(otherwise heudiconv will crash when trying to create a file which
already exists). A rule has been proposed: leave the run which is
longest and latest (i. e. sort runs first by dim4 and then by
series_id). This assumes that series_id starts with a number,
which should be the case.

This should mitigate a special case of issue #1