Fix sorting of repeated series to take number from series_id
Our series_ids appear to have the form: N-protocol_name.
Previously series_id was sorted as strings, so 9-task-run would
come after 10-task-run. Now sorting is done on the integers
extracted from the beginning of the series id (the sorting is
done only when task-<>_run-<> repeats in the study).

This assumes that the series id will indeed start with a number.